pr review and test cases
